+++ src/data/systems/broken-compass/components/create-character/BCCharacterCreateGeneral.svelte
@@ -14,14 +14,14 @@
   </label>
   <label class="label">
     <span>Heritage</span>
-    <input class="input" type="text" placeholder="Heritage" bind:value={characterHeritage} />
+    <input class="input" type="text" placeholder="Where you were born" bind:value={characterHeritage} />
   </label>
   <label class="label">
     <span>Homeland</span>
-    <input class="input" type="text" placeholder="Homeland" bind:value={characterHomeland} />
+    <input class="input" type="text" placeholder="Where you grew up" bind:value={characterHomeland} />
   </label>
   <label class="label">
     <span>Workplace</span>
-    <input class="input" type="text" placeholder="Workplace" bind:value={characterWorkplace} />
+    <input class="input" type="text" placeholder="Where you do you business" bind:value={characterWorkplace} />
   </label>
 </form>